# openapi.yaml
openapi: 3.0.3
info:
  title: Digital Notary MVP API
  version: 0.1.0
paths:
...
  /v1/documents:
    post:
      summary: Create document after accepting the EDO agreement
      security: [{bearerAuth: []}]
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/CreateDocument'}}}
      responses: {'200': {description: Created}}
    get: {summary: List the current user's documents, parameters: [{name: limit, in: query, schema: {type: integer, maximum: 100}}], responses: {'200': {description: Documents}}}
...
components:
  securitySchemes: {bearerAuth: {type: http, scheme: bearer}}
  schemas:
    CreateDocument:
      type: object
      required: [title, contractorPhone, content, edoAgreementVersion]
      properties:
        title: {type: string}
        template: {type: string, example: gph-contract}
        contractorPhone: {type: string, example: '+79990000000'}
        content: {type: string, description: Document body or immutable uploaded-file checksum input}
        edoAgreementVersion: {type: string, example: '2026-08-01'}
        amountKopecks: {type: integer, format: int64}